created abstract car class for use with all cars
