+++ src/main/java/jMonkeyEngine/Entities/Cars/VehicleModels/Vehicle.java
@@ -35,15 +35,15 @@ abstract public class Vehicle
     // constants and loggers
 
     /**
-     * factor to convert km/hr to miles per hour
+     * factor to convert km/hr to miles per hour.
      */
     final public static float KPH_TO_MPH = 0.62137f;
     /**
-     * factor to convert km/hr to wu/sec
+     * factor to convert km/hr to wu/sec.
      */
     final public static float KPH_TO_WUPS = 0.277778f;
     /**
-     * message logger for this class
+     * message logger for this class.
      */
     final public static Logger logger
             = Logger.getLogger(Vehicle.class.getName());
@@ -51,52 +51,51 @@ abstract public class Vehicle
     // fields
 
     /**
-     * linear damping due to air resistance on the chassis (&ge;0, &lt;1)
+     * linear damping due to air resistance on the chassis (&ge;0, &lt;1).
      */
     private float chassisDamping;
     /**
-     * the fraction of the total mass in each body (each element &ge;0, &le;1)
+     * the fraction of the total mass in each body (each element &ge;0, &le;1).
      * or null if not determined yet
      */
     private float[] massFractions;
     /**
-     * ratio of the steeringWheelAngle to the turn angle of any wheels used for
-     * steering
+     * ratio of the steeringWheelAngle to the turn angle of any wheels used for steering.
      */
     private float steeringRatio = 2f;
     /**
      * rotation of the steering wheel, handlebars, or tiller (in radians,
-     * negative&rarr;left, 0&rarr;neutral, positive&rarr;right)
+     * negative&rarr;left, 0&rarr;neutral, positive&rarr;right).
      */
     private float steeringWheelAngle;
     /**
-     * physics body associated with the Engine
+     * physics body associated with the Engine.
      */
     private VehicleControl engineBody;
     /**
-     * support the chassis and configure acceleration, steering, and braking
+     * support the chassis and configure acceleration, steering, and braking.
      */
     final private List<Wheel> wheels = new ArrayList<>(4);
     /**
-     * temporary storage for the vehicle's orientation
+     * temporary storage for the vehicle's orientation.
      */
     final private static Matrix3f tmpOrientation = new Matrix3f();
     /**
-     * scene-graph subtree that represents this Vehicle
+     * scene-graph subtree that represents this Vehicle.
      */
     final private Node node;
     /**
      * computer-graphics (C-G) model to visualize the whole Vehicle except for
-     * its wheels
+     * its wheels.
      */
     private Spatial chassis;
     /**
-     * descriptive name (not null)
+     * descriptive name (not null).
      */
     final private String name;
     /**
      * default transform of each body relative to the engine body, or null if
-     * transforms have not yet been determined
+     * transforms have not yet been determined.
      */
     private Transform[] relativeTransforms;
     // *************************************************************************
